refactor: Ladder 클래스 분리
