+++ src/main/java/ladder/LadderGame.java
@@ -1,0 +1,17 @@
+package ladder;
+
+import ladder.creator.LadderCreator;
+
+public class LadderGame {
+
+    private final LadderCreator ladderCreator;
+
+    public LadderGame(LadderCreator ladderCreator) {
+        this.ladderCreator = ladderCreator;
+    }
+
+    public int run(Position position) {
+        LadderRunner ladderRunner = new LadderRunner(ladderCreator.getRows());
+        return ladderRunner.run(position).getPosition();
+    }
+}
